update user as soon as created
